2311 Merge WebhookWorkflowExecutionFacade into WebhookWorkflowExecutor
