+++ server/libs/platform/platform-webhook/platform-webhook-api/src/main/java/com/bytechef/platform/webhook/executor/WebhookWorkflowExecutor.java
@@ -17,6 +17,7 @@
 package com.bytechef.platform.webhook.executor;
 
 import com.bytechef.component.definition.TriggerDefinition.WebhookValidateResponse;
+import com.bytechef.platform.component.domain.WebhookTriggerFlags;
 import com.bytechef.platform.component.trigger.WebhookRequest;
 import com.bytechef.platform.job.sync.SseStreamBridge;
 import com.bytechef.platform.workflow.WorkflowExecutionId;
@@ -27,6 +28,12 @@ import org.jspecify.annotations.Nullable;
  * Defines the contract for executing and managing webhook-based workflow executions. Provides methods to execute
  * workflows asynchronously, synchronously, or with SSE streaming support. Additionally, includes validation
  * capabilities for webhook requests during execution or enablement processes.
+ *
+ * <p>
+ * Also exposes the transport-agnostic orchestration helpers (formerly the {@code WebhookWorkflowExecutionFacade}) that
+ * callers across transports (HTTP/SSE controller, AI copilot webhook bridge) compose to drive a webhook-triggered run:
+ * {@link #executeStreaming}, {@link #getWebhookTriggerFlags}, {@link #hasStreamingTask}, and
+ * {@link #isWorkflowDisabled}.
  *
  * @author Ivica Cardic
  */
@@ -73,6 +80,50 @@ public interface WebhookWorkflowExecutor {
      */
     @Nullable
     Object executeSync(WorkflowExecutionId workflowExecutionId, WebhookRequest webhookRequest);
+
+    /**
+     * Executes a workflow with streaming output. Events are pushed to {@code sseStreamBridge} as the workflow runs;
+     * completion / error are signalled via {@link SseStreamBridge#onComplete()} / {@link SseStreamBridge#onError}.
+     *
+     * <p>
+     * When the workflow is disabled the bridge receives a synchronous {@link SseStreamBridge#onError} call with an
+     * {@code IllegalStateException}, mirroring the HTTP controller's pre-existing behaviour. The returned future
+     * completes immediately in that case.
+     * </p>
+     *
+     * @param workflowExecutionId the workflow execution to invoke
+     * @param webhookRequest      the request payload
+     * @param sseStreamBridge     the transport-specific event sink (SSE emitter wrap, AG-UI bridge, etc.)
+     * @return a future that completes when the workflow run finishes or errors
+     */
+    CompletableFuture<Void> executeStreaming(
+        WorkflowExecutionId workflowExecutionId, WebhookRequest webhookRequest, SseStreamBridge sseStreamBridge);
+
+    /**
+     * Returns the trigger flags associated with the workflow's webhook trigger. Cached lookup — the underlying
+     * {@code TriggerDefinitionService} call is the same one the controller already uses to decide between sync and
+     * async dispatch on the HTTP path. The copilot bridge consumes this to pick between
+     * {@link #executeSync(WorkflowExecutionId, WebhookRequest)} and
+     * {@link #executeStreaming(WorkflowExecutionId, WebhookRequest, SseStreamBridge)}.
+     */
+    WebhookTriggerFlags getWebhookTriggerFlags(WorkflowExecutionId workflowExecutionId);
+
+    /**
+     * Returns {@code true} when the workflow contains at least one task whose action streams output (e.g.
+     * {@code openAi/v1/streamAsk}, {@code aiAgent/v1/streamChat}). The copilot bridge consults this flag BEFORE looking
+     * at the trigger's {@code workflowSyncExecution} property: a chat workflow whose trigger demands sync semantics for
+     * HTTP transport (so embedded chat clients keep working) but contains a streaming AI task needs the bridge to route
+     * through {@link #executeStreaming} so per-token deltas flow to the AG-UI client. Otherwise the streaming task's
+     * events would land in JobSyncExecutor's in-process bridge cache — unreachable from EE worker processes — and the
+     * user would see a blank assistant reply.
+     */
+    boolean hasStreamingTask(WorkflowExecutionId workflowExecutionId);
+
+    /**
+     * Checks whether the workflow associated with the given execution id is currently disabled. Both transports use
+     * this to short-circuit the dispatch path before execution is invoked.
+     */
+    boolean isWorkflowDisabled(WorkflowExecutionId workflowExecutionId);
 
     /**
      * Validates the provided webhook request and executes the associated workflow asynchronously based on the given
