+++ server/libs/platform/platform-webhook/platform-webhook-rest/platform-webhook-rest-impl/src/main/java/com/bytechef/platform/webhook/web/rest/WebhookTriggerController.java
@@ -16,21 +16,17 @@
 
 package com.bytechef.platform.webhook.web.rest;
 
-import com.bytechef.atlas.configuration.service.WorkflowService;
 import com.bytechef.atlas.coordinator.annotation.ConditionalOnCoordinator;
 import com.bytechef.commons.util.JsonUtils;
 import com.bytechef.component.definition.TriggerDefinition.WebhookValidateResponse;
 import com.bytechef.config.ApplicationProperties;
 import com.bytechef.platform.component.domain.WebhookTriggerFlags;
-import com.bytechef.platform.component.service.TriggerDefinitionService;
 import com.bytechef.platform.component.trigger.WebhookRequest;
 import com.bytechef.platform.file.storage.TempFileStorage;
 import com.bytechef.platform.job.sync.SseStreamBridge;
-import com.bytechef.platform.webhook.executor.WebhookWorkflowExecutionFacade;
 import com.bytechef.platform.webhook.executor.WebhookWorkflowExecutor;
 import com.bytechef.platform.webhook.rest.AbstractWebhookTriggerController;
 import com.bytechef.platform.workflow.WorkflowExecutionId;
-import com.bytechef.platform.workflow.execution.accessor.JobPrincipalAccessorRegistry;
 import com.bytechef.tenant.TenantContext;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import jakarta.servlet.http.HttpServletRequest;
@@ -65,21 +61,15 @@ public class WebhookTriggerController extends AbstractWebhookTriggerController {
 
     private static final Logger log = LoggerFactory.getLogger(WebhookTriggerController.class);
 
-    private final WebhookWorkflowExecutionFacade webhookWorkflowExecutionFacade;
     private final WebhookWorkflowExecutor webhookWorkflowExecutor;
 
     @SuppressFBWarnings("EI")
     public WebhookTriggerController(
-        ApplicationProperties applicationProperties, JobPrincipalAccessorRegistry jobPrincipalAccessorRegistry,
-        TempFileStorage tempFileStorage, TriggerDefinitionService triggerDefinitionService,
-        WebhookWorkflowExecutionFacade webhookWorkflowExecutionFacade, WebhookWorkflowExecutor webhookWorkflowExecutor,
-        WorkflowService workflowService) {
+        ApplicationProperties applicationProperties, TempFileStorage tempFileStorage,
+        WebhookWorkflowExecutor webhookWorkflowExecutor) {
 
-        super(
-            jobPrincipalAccessorRegistry, applicationProperties.getPublicUrl(), tempFileStorage,
-            triggerDefinitionService, webhookWorkflowExecutor, workflowService);
+        super(applicationProperties.getPublicUrl(), tempFileStorage, webhookWorkflowExecutor);
 
-        this.webhookWorkflowExecutionFacade = webhookWorkflowExecutionFacade;
         this.webhookWorkflowExecutor = webhookWorkflowExecutor;
     }
 
@@ -112,11 +102,11 @@ public class WebhookTriggerController extends AbstractWebhookTriggerController {
             ResponseEntity<?> responseEntity;
 
             boolean head = Objects.equals(httpServletRequest.getMethod(), RequestMethod.HEAD.name());
-            boolean disabled = webhookWorkflowExecutionFacade.isWorkflowDisabled(workflowExecutionId);
+            boolean disabled = webhookWorkflowExecutor.isWorkflowDisabled(workflowExecutionId);
 
             if (head || disabled) {
                 WebhookTriggerFlags webhookTriggerFlags =
-                    webhookWorkflowExecutionFacade.getWebhookTriggerFlags(workflowExecutionId);
+                    webhookWorkflowExecutor.getWebhookTriggerFlags(workflowExecutionId);
 
                 WebhookRequest webhookRequest = getWebhookRequest(httpServletRequest, webhookTriggerFlags);
 
@@ -166,10 +156,11 @@ public class WebhookTriggerController extends AbstractWebhookTriggerController {
         return TenantContext.callWithTenantId(workflowExecutionId.getTenantId(), () -> {
             WebhookSseStreamBridge bridge = new WebhookSseStreamBridge(emitter);
 
-            WebhookTriggerFlags webhookTriggerFlags = getWebhookTriggerFlags(workflowExecutionId);
+            WebhookTriggerFlags webhookTriggerFlags = webhookWorkflowExecutor.getWebhookTriggerFlags(
+                workflowExecutionId);
             WebhookRequest webhookRequest = getWebhookRequest(httpServletRequest, webhookTriggerFlags);
 
-            webhookWorkflowExecutionFacade.executeStreaming(workflowExecutionId, webhookRequest, bridge);
+            webhookWorkflowExecutor.executeStreaming(workflowExecutionId, webhookRequest, bridge);
 
             return emitter;
         });
